requirements.txt
Click==7.0
Flask==1.0.2
gunicorn==19.9.0
image-slicer==0.2.0
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
Pillow==5.4.1
Werkzeug==0.14.1
